Simple yet powerful Messaging application for  Android
Efficient messaging application that takes much less time in loading and
displaying all the messages in the device.
It has many more features than native Android messaging app like swipe
to delete.
More features being added.

- It provides a way to send and receive SMSs.
- It displays the messages you have received in a format specified
below:

- For a given sender, UI displays only one tile showing the latest
message received for that sender. If this message is unread, it should
be displayed in bold.
- A tile can be colored in two ways based on two situations that can
arise, the sender is in your phonebook or it is an unknown number.
- Tiles are swipable and exhibit different behavior based on its color.
If it is an unknown contact, the tile gets deleted. If it is from a
contact from your phonebook, the message is just marked as read i.e. the
text will not be bold now.
- Shows a notification once you receive an SMS.
- The application is expected to run smoothly and should be performant.